pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "ctf-destroyer-poc"
version = "0.1.0"
description = "LangGraph PoC orchestrating Codex and Claude subprocess workers for CTF challenges."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "langgraph>=1.0.10",
  "PyYAML>=6.0.2",
]

[project.scripts]
ctf-orchestrator = "ctf_destroyer.cli:main"
ctf-import = "ctf_destroyer.import_cli:main"

[tool.setuptools.packages.find]
include = ["ctf_destroyer*"]
